+++ gem/spec/generators/generator_helper.rb
@@ -22,6 +22,7 @@ require "tmpdir"
 
 require_relative "../../lib/generators/ecs_rails/install/install_generator"
 require_relative "../../lib/generators/ecs_rails/component/component_generator"
+require_relative "../../lib/generators/ecs_rails/relationship/relationship_generator"
 
 # Shared behaviour for generator examples. Include with `include GeneratorHelper`
 # or rely on the `type: :generator` metadata hook configured below.